+++ lib/public/views/Runs/Overview/RunsOverviewModel.js
@@ -86,8 +86,6 @@ export class RunsOverviewModel extends OverviewPageModel {
         // Export items
         this._allRuns = RemoteData.NotAsked();
 
-        this._pdpBeamType = null;
-
         this.reset(false);
         // eslint-disable-next-line no-return-assign,require-jsdoc
         const updateDebounceTime = () => this._debouncedLoad = debounce(this.load.bind(this), model.inputDebounceTime);
@@ -110,7 +108,7 @@ export class RunsOverviewModel extends OverviewPageModel {
      */
     async load() {
         this._allRuns = RemoteData.NotAsked();
-        await super.load();
+        super.load();
     }
 
     /**
@@ -169,7 +167,6 @@ export class RunsOverviewModel extends OverviewPageModel {
      */
     reset(fetch = true) {
         super.reset();
-        this._pdpBeamType = new Set();
         this.resetFiltering(fetch);
     }
 
@@ -835,24 +832,6 @@ export class RunsOverviewModel extends OverviewPageModel {
      */
     get aliceL3AndDipoleCurrentFilter() {
         return this._aliceL3AndDipoleCurrentFilter;
-    }
-
-    /**
-     * Get pdp_beam_type of fetched runs
-     *
-     * @return {string} beam type
-     */
-    get pdpBeamTypes() {
-        return this._pdpBeamType;
-    }
-
-    /**
-     * Set pdp_beam_type of fetched runs
-     *
-     * @param {string} pdpBeamType beam type
-     */
-    set pdpBeamTypes(pdpBeamType) {
-        this._pdpBeamType = pdpBeamType;
     }
 
     /**
