+++ lib/public/views/Runs/RunPerDataPass/RunsPerDataPassOverviewModel.js
@@ -12,16 +12,16 @@
  */
 import { RemoteData } from '/js/src/index.js';
 import { buildUrl } from '../../../utilities/fetch/buildUrl.js';
-import { RunsOverviewModel } from '../Overview/RunsOverviewModel.js';
 import { ObservableData } from '../../../utilities/ObservableData.js';
 import { getRemoteDataSlice } from '../../../utilities/fetch/getRemoteDataSlice.js';
 import { getRemoteData } from '../../../utilities/fetch/getRemoteData.js';
 import { detectorsProvider } from '../../../services/detectors/detectorsProvider.js';
+import { FixedPdpBeamTypeRunsOverviewModel } from '../Overview/FixedPdpBeamTypeRunsOverviewModel.js';
 
 /**
  * Runs Per Data Pass overview model
  */
-export class RunsPerDataPassOverviewModel extends RunsOverviewModel {
+export class RunsPerDataPassOverviewModel extends FixedPdpBeamTypeRunsOverviewModel {
     /**
      * Constructor
      * @param {Model} model global model
