+++ lib/public/views/Runs/RunPerDataPass/RunsPerDataPassOverviewPage.js
@@ -77,8 +77,8 @@ export const RunsPerDataPassOverviewPage = ({
             pp: inelasticInteractionRateActiveColumnsForProtonProton,
             PbPb: inelasticInteractionRateActiveColumnsForPbPb,
         }, {
-            ...inelasticInteractionRateActiveColumnsForPbPb,
             ...inelasticInteractionRateActiveColumnsForProtonProton,
+            ...inelasticInteractionRateActiveColumnsForPbPb,
         }),
         ...remoteGaqSummary.match({
             Success: (gaqSummary) => ({
